Refactor channel grid and layouts, add TV support
