+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/dev/Main_Layout.kt
@@ -1,46 +1,44 @@
 package com.skylake.skytv.jgorunner.ui.dev
 
-import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.util.Log
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.core.content.ContextCompat.startActivity
-import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
-import com.bumptech.glide.integration.compose.GlideImage
-import com.skylake.skytv.jgorunner.data.SkySharedPref
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.focus.onFocusEvent
-import androidx.compose.ui.layout.ContentScale
-import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Done
 import androidx.compose.material.icons.filled.Refresh
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.core.content.ContextCompat.startActivity
+import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
+import com.bumptech.glide.integration.compose.GlideImage
+import com.skylake.skytv.jgorunner.activities.ChannelInfo
+import com.skylake.skytv.jgorunner.data.SkySharedPref
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.ElevatedCard
+import androidx.compose.material3.FilterChip
+import androidx.compose.material3.FilterChipDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
-import com.skylake.skytv.jgorunner.activities.ChannelInfo
 import com.skylake.skytv.jgorunner.services.player.ExoPlayJet
 import com.skylake.skytv.jgorunner.ui.screens.AppStartTracker
 
-@SuppressLint("MutableCollectionMutableState")
 @OptIn(ExperimentalGlideComposeApi::class)
 @Composable
 fun Main_Layout(context: Context?) {
@@ -50,7 +48,7 @@ fun Main_Layout(context: Context?) {
     val filteredChannels = remember { mutableStateOf<List<Channel>>(emptyList()) }
     val preferenceManager = SkySharedPref.getInstance(context)
     val localPORT by remember {
-        mutableIntStateOf(preferenceManager.myPrefs.jtvGoServerPort)
+        mutableIntStateOf(preferenceManager.myPrefs.jtvGoServerPort ?: 8080)
     }
     val basefinURL = "http://localhost:$localPORT"
     var fetched by remember { mutableStateOf(false) }
@@ -59,7 +57,6 @@ fun Main_Layout(context: Context?) {
     var epgData by remember { mutableStateOf<EpgProgram?>(null) }
 
     val focusRequester = remember { FocusRequester() }
-
     val categoryMap = mapOf(
         "Reset" to null,
         "Entertainment" to 5,
@@ -77,11 +74,9 @@ fun Main_Layout(context: Context?) {
         "JioDarshan" to 19
     )
 
-    val categoryIdMap = categoryMap.filterValues { it != null }.map { it.value!! to it.key }.toMap()
     val savedCategoryIds = preferenceManager.myPrefs.filterCI
         ?.split(",")?.mapNotNull { it.toIntOrNull() }?.toMutableSet() ?: mutableSetOf()
     var selectedCategoryIds by remember { mutableStateOf(savedCategoryIds) }
-    val selectedCategories = categoryMap.filterValues { it in selectedCategoryIds }.keys
 
     val sortedCategories = remember(selectedCategoryIds) {
         val resetCategoryName = "Reset"
@@ -94,13 +89,14 @@ fun Main_Layout(context: Context?) {
         listOf(resetCategoryName) + selectedOtherCategories + unselectedOtherCategories
     }
 
+    // Fetch and filter channels (cache/network)
     LaunchedEffect(Unit) {
         val sharedPref = context.getSharedPreferences("channel_cache", Context.MODE_PRIVATE)
         var useCache = true
         var cachedChannels: ChannelResponse? = null
 
         val cachedJson = sharedPref.getString("channels_json", null)
-        if (cachedJson != null) {
+        if (!cachedJson.isNullOrEmpty()) {
             try {
                 cachedChannels = Gson().fromJson(cachedJson, ChannelResponse::class.java)
                 channelsResponse.value = cachedChannels
@@ -120,36 +116,22 @@ fun Main_Layout(context: Context?) {
                 ?.split(",")?.map { it.trim() }?.filter { it.isNotEmpty() }
             val languages = preferenceManager.myPrefs.filterLI
                 ?.split(",")?.map { it.trim() }?.filter { it.isNotEmpty() }
+
             val filtered = when {
-                categories.isNullOrEmpty() && languages.isNullOrEmpty() -> {
-                    ChannelUtils.filterChannels(cachedChannels)
-                }
-
-                categories.isNullOrEmpty() -> {
-                    ChannelUtils.filterChannels(
-                        cachedChannels,
-                        languageIds = languages?.mapNotNull { it.toIntOrNull() }
-                            ?.takeIf { it.isNotEmpty() }
-                    )
-                }
-
-                languages.isNullOrEmpty() -> {
-                    ChannelUtils.filterChannels(
-                        cachedChannels,
-                        categoryIds = categories.mapNotNull { it.toIntOrNull() }
-                            .takeIf { it.isNotEmpty() }
-                    )
-                }
-
-                else -> {
-                    ChannelUtils.filterChannels(
-                        cachedChannels,
-                        categoryIds = categories.mapNotNull { it.toIntOrNull() }
-                            .takeIf { it.isNotEmpty() },
-                        languageIds = languages.mapNotNull { it.toIntOrNull() }
-                            .takeIf { it.isNotEmpty() }
-                    )
-                }
+                categories.isNullOrEmpty() && languages.isNullOrEmpty() -> ChannelUtils.filterChannels(cachedChannels)
+                categories.isNullOrEmpty() -> ChannelUtils.filterChannels(
+                    cachedChannels,
+                    languageIds = languages?.mapNotNull { it.toIntOrNull() }?.takeIf { it.isNotEmpty() }
+                )
+                languages.isNullOrEmpty() -> ChannelUtils.filterChannels(
+                    cachedChannels,
+                    categoryIds = categories.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() }
+                )
+                else -> ChannelUtils.filterChannels(
+                    cachedChannels,
+                    categoryIds = categories.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() },
+                    languageIds = languages.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() }
+                )
             }
             filteredChannels.value = filtered ?: emptyList()
             fetched = true
@@ -161,7 +143,6 @@ fun Main_Layout(context: Context?) {
                 try {
                     val response = ChannelUtils.fetchChannels("$basefinURL/channels")
                     channelsResponse.value = response
-
                     if (response != null) {
                         val responseJsonString = Gson().toJson(response)
                         with(sharedPref.edit()) {
@@ -172,41 +153,28 @@ fun Main_Layout(context: Context?) {
                             ?.split(",")?.map { it.trim() }?.filter { it.isNotEmpty() }
                         val languages = preferenceManager.myPrefs.filterLI
                             ?.split(",")?.map { it.trim() }?.filter { it.isNotEmpty() }
+
                         val filtered = when {
-                            categories.isNullOrEmpty() && languages.isNullOrEmpty() -> {
-                                ChannelUtils.filterChannels(response)
-                            }
-
-                            categories.isNullOrEmpty() -> {
-                                ChannelUtils.filterChannels(
-                                    response,
-                                    languageIds = languages?.mapNotNull { it.toIntOrNull() }
-                                        ?.takeIf { it.isNotEmpty() }
-                                )
-                            }
-
-                            languages.isNullOrEmpty() -> {
-                                ChannelUtils.filterChannels(
-                                    response,
-                                    categoryIds = categories.mapNotNull { it.toIntOrNull() }
-                                        ?.takeIf { it.isNotEmpty() }
-                                )
-                            }
-
-                            else -> {
-                                ChannelUtils.filterChannels(
-                                    response,
-                                    categoryIds = categories.mapNotNull { it.toIntOrNull() }
-                                        ?.takeIf { it.isNotEmpty() },
-                                    languageIds = languages.mapNotNull { it.toIntOrNull() }
-                                        ?.takeIf { it.isNotEmpty() }
-                                )
-                            }
+                            categories.isNullOrEmpty() && languages.isNullOrEmpty() -> ChannelUtils.filterChannels(response)
+                            categories.isNullOrEmpty() -> ChannelUtils.filterChannels(
+                                response,
+                                languageIds = languages?.mapNotNull { it.toIntOrNull() }?.takeIf { it.isNotEmpty() }
+                            )
+                            languages.isNullOrEmpty() -> ChannelUtils.filterChannels(
+                                response,
+                                categoryIds = categories.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() }
+                            )
+                            else -> ChannelUtils.filterChannels(
+                                response,
+                                categoryIds = categories.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() },
+                                languageIds = languages.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() }
+                            )
                         }
                         filteredChannels.value = filtered ?: emptyList()
                         success = true
                     }
-                } catch (_: Exception) {
+                } catch (e: Exception) {
+                    // ignore, retry
                 }
                 if (!success) {
                     kotlinx.coroutines.delay(300)
@@ -214,9 +182,6 @@ fun Main_Layout(context: Context?) {
             }
             fetched = true
         }
-
-
-
 
         if (preferenceManager.myPrefs.startTvAutomatically) {
             if (!AppStartTracker.shouldPlayChannel &&
@@ -243,7 +208,6 @@ fun Main_Layout(context: Context?) {
                     putExtra("ch_name", firstChannel.channel_name ?: "")
                 }
                 kotlinx.coroutines.delay(1000)
-
                 startActivity(context, intent, null)
 
                 //Recent Channels
@@ -271,12 +235,10 @@ fun Main_Layout(context: Context?) {
 
             AppStartTracker.shouldPlayChannel = true
         }
+
     }
 
-
-
-
-
+    // Re-filter channels when category changes
     LaunchedEffect(selectedCategoryIds) {
         channelsResponse.value?.let { response ->
             val languages = preferenceManager.myPrefs.filterLI
@@ -291,17 +253,18 @@ fun Main_Layout(context: Context?) {
         }
     }
 
+    // Fetch EPG data for selected channel
     LaunchedEffect(selectedChannel) {
         selectedChannel?.let { channel ->
-            val epgURLc = "$basefinURL/epg/${channel.channel_id}/0"
+            val epgURLc = "$basefinURL/epg/${channel.channel_id ?: ""}/0"
             Log.d("NANOdix",epgURLc)
             epgData = ChannelUtils.fetchEpg(epgURLc)
-            Log.d("NANOdix", "Now playing: ${epgData?.showname}")
         } ?: run {
             epgData = null
         }
     }
 
+    // UI: Loading state
     if (!fetched) {
         Box(
             contentAlignment = Alignment.Center,
@@ -309,7 +272,9 @@ fun Main_Layout(context: Context?) {
         ) {
             CircularProgressIndicator(modifier = Modifier.size(60.dp))
         }
-    } else if (filteredChannels.value.isNullOrEmpty()) {
+    }
+    // UI: Empty state
+    else if (filteredChannels.value.isNullOrEmpty()) {
         Box(
             contentAlignment = Alignment.Center,
             modifier = Modifier.fillMaxSize()
@@ -317,18 +282,18 @@ fun Main_Layout(context: Context?) {
             Column(horizontalAlignment = Alignment.CenterHorizontally) {
                 Text(
                     text = "No channels found",
-                    style = MaterialTheme.typography.titleLarge,
+                    style = TextStyle(fontSize = 20.sp, fontWeight = FontWeight.Bold),
                     color = Color.Red
                 )
                 Spacer(modifier = Modifier.height(16.dp))
                 Text(
                     text = "Try the following steps:",
-                    style = MaterialTheme.typography.bodyLarge
+                    style = TextStyle(fontSize = 16.sp)
                 )
                 Spacer(modifier = Modifier.height(8.dp))
                 Text(
                     text = "\n• Check your internet connection\n• Reload the app\n• Go to Main Screen for detailed information",
-                    style = MaterialTheme.typography.bodyMedium.copy(color = Color.Gray)
+                    style = TextStyle(fontSize = 15.sp, color = Color.Gray)
                 )
                 Spacer(modifier = Modifier.height(24.dp))
                 ElevatedCard(
@@ -345,14 +310,19 @@ fun Main_Layout(context: Context?) {
                         verticalAlignment = Alignment.CenterVertically,
                         modifier = Modifier.padding(horizontal = 24.dp, vertical = 12.dp)
                     ) {
-                        Icon(imageVector = Icons.Filled.Refresh, contentDescription = "Reload")
+                        Icon(
+                            imageVector = Icons.Filled.Refresh,
+                            contentDescription = "Reload"
+                        )
                         Spacer(modifier = Modifier.width(8.dp))
                         Text("Reload App")
                     }
                 }
             }
         }
-    } else {
+    }
+    // UI: Main content
+    else {
         // CATEGORY CHIPS
         LazyRow(
             modifier = Modifier
@@ -363,6 +333,7 @@ fun Main_Layout(context: Context?) {
             items(sortedCategories) { categoryName ->
                 val categoryId = categoryMap[categoryName]
                 val isSelected = categoryId != null && selectedCategoryIds.contains(categoryId)
+
                 FilterChip(
                     onClick = {
                         if (categoryName == "Reset") {
@@ -379,7 +350,7 @@ fun Main_Layout(context: Context?) {
                         preferenceManager.myPrefs.filterCI = updatedCI
                         preferenceManager.savePreferences()
                     },
-                    label = { Text(categoryName) },
+                    label = { Text(categoryName ?: "Unknown") },
                     selected = isSelected,
                     leadingIcon = when {
                         categoryName == "Reset" -> {
@@ -406,7 +377,7 @@ fun Main_Layout(context: Context?) {
             }
         }
 
-        // EPG CARD
+        // EPG CARD (null-safe)
         epgData?.let { epg ->
             Column(
                 modifier = Modifier
@@ -431,23 +402,22 @@ fun Main_Layout(context: Context?) {
                         ) {
                             Text(
                                 text = epg.channel_name ?: "Unknown Channel",
-                                style = MaterialTheme.typography.bodyMedium
+                                style = TextStyle(fontSize = 14.sp)
                             )
                             Spacer(modifier = Modifier.height(6.dp))
                             Text(
                                 text = epg.showname ?: "No Program Info",
                                 maxLines = 1,
-                                style = MaterialTheme.typography.titleLarge.copy(fontWeight = FontWeight.Bold)
+                                style = TextStyle(fontSize = 22.sp, fontWeight = FontWeight.Bold)
                             )
                             Spacer(modifier = Modifier.height(8.dp))
                             Text(
                                 text = epg.description ?: "No description available",
-                                style = MaterialTheme.typography.bodySmall,
-                                maxLines = 2,
+                                style = TextStyle(fontSize = 13.sp),
+                                maxLines = 3,
                                 overflow = TextOverflow.Ellipsis
                             )
                         }
-
                         epg.episodePoster?.let { poster ->
                             GlideImage(
                                 model = "$basefinURL/jtvposter/$poster",
@@ -463,95 +433,101 @@ fun Main_Layout(context: Context?) {
             }
         }
 
+        ChannelGridMain(
+            context = context,
+            filteredChannels = filteredChannels.value ?: emptyList(),
+            selectedChannelSetter = { selectedChannel = it },
+            localPORT = localPORT,
+            preferenceManager = preferenceManager
+        )
+
+
         // CHANNEL GRID
-        LazyVerticalGrid(
-            columns = GridCells.Adaptive(minSize = 100.dp),
-            contentPadding = PaddingValues(16.dp),
-            verticalArrangement = Arrangement.spacedBy(16.dp),
-            horizontalArrangement = Arrangement.spacedBy(16.dp)
-        ) {
-            items(filteredChannels.value ?: emptyList()) { channel ->
-                var isFocused by remember { mutableStateOf(false) }
-                ElevatedCard(
-                    elevation = CardDefaults.cardElevation(defaultElevation = 6.dp),
-                    modifier = Modifier
-                        .height(120.dp)
-                        .focusRequester(focusRequester)
-                        .onFocusEvent { focusState ->
-                            isFocused = focusState.isFocused
-                            if (focusState.isFocused) {
-                                selectedChannel = channel
-                            }
-                        }
-                        .clickable {
-                            val intent = Intent(context, ExoPlayJet::class.java).apply {
-                                putExtra("zone", "TV")
-                                putParcelableArrayListExtra("channel_list_data", ArrayList(
-                                    (filteredChannels.value ?: emptyList()).map { ch ->
-                                        ChannelInfo(
-                                            ch.channel_url ?: "",
-                                            "http://localhost:$localPORT/jtvimage/${ch.logoUrl ?: ""}",
-                                            ch.channel_name ?: ""
-                                        )
-                                    }
-                                ))
-                                putExtra("current_channel_index", (filteredChannels.value ?: emptyList()).indexOf(channel))
-                                putExtra("video_url", channel.channel_url ?: "")
-                                putExtra("logo_url", "http://localhost:$localPORT/jtvimage/${channel.logoUrl ?: ""}")
-                                putExtra("ch_name", channel.channel_name ?: "")
-                            }
-                            startActivity(context, intent, null)
-
-                            // Update recent channels
-                            val recentChannelsJson = preferenceManager.myPrefs.recentChannels
-                            val type = object : TypeToken<List<Channel>>() {}.type
-                            val recentChannels: MutableList<Channel> =
-                                Gson().fromJson(recentChannelsJson, type) ?: mutableListOf()
-
-                            val existingIndex = recentChannels.indexOfFirst {
-                                it.channel_id == channel.channel_id
-                            }
-
-                            if (existingIndex != -1) {
-                                val existingChannel = recentChannels[existingIndex]
-                                recentChannels.removeAt(existingIndex)
-                                recentChannels.add(0, existingChannel)
-                            } else {
-                                recentChannels.add(0, channel)
-                                if (recentChannels.size > 25) {
-                                    recentChannels.removeAt(recentChannels.size - 1)
-                                }
-                            }
-
-                            preferenceManager.myPrefs.recentChannels = Gson().toJson(recentChannels)
-                            preferenceManager.savePreferences()
-                        },
-                    colors = CardDefaults.cardColors(
-                        containerColor = MaterialTheme.colorScheme.primaryContainer
-                    )
-                ) {
-                    Column {
-                        // Channel logo
-                        val imageUrl = "http://localhost:$localPORT/jtvimage/${channel.logoUrl ?: ""}"
-
-                        GlideImage(
-                            model = imageUrl,
-                            contentDescription = channel.channel_name ?: "",
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .height(80.dp),
-                            contentScale = ContentScale.Fit
-                        )
-
-                        // Channel name
-                        Text(
-                            text = channel.channel_name ?: "Unknown",
-                            fontSize = 12.sp,
-                            modifier = Modifier.padding(8.dp)
-                        )
-                    }
-                }
-            }
-        }
+//        LazyVerticalGrid(
+//            columns = GridCells.Adaptive(minSize = 100.dp),
+//            contentPadding = PaddingValues(16.dp),
+//            verticalArrangement = Arrangement.spacedBy(16.dp),
+//            horizontalArrangement = Arrangement.spacedBy(16.dp)
+//        ) {
+//            items(filteredChannels.value ?: emptyList()) { channel ->
+//                if (channel == null) return@items // Defensive: skip nulls
+//                ClassicCard(
+//                    modifier = Modifier
+//                        .height(120.dp)
+//                        .focusRequester(focusRequester)
+//                        .onFocusEvent { focusState ->
+//                            if (focusState.isFocused) {
+//                                selectedChannel = channel
+//                            }
+//                        },
+//                    image = {
+//                        val logoUrl = channel.logoUrl ?: ""
+//                        val imageUrl = "$basefinURL/jtvimage/$logoUrl"
+//                        GlideImage(
+//                            model = imageUrl,
+//                            contentDescription = channel.channel_name ?: "Channel",
+//                            modifier = Modifier
+//                                .fillMaxWidth()
+//                                .height(80.dp)
+//                        )
+//                    },
+//                    title = {
+//                        Text(
+//                            text = channel.channel_name ?: "Channel",
+//                            fontSize = 12.sp,
+//                            modifier = Modifier.padding(8.dp)
+//                        )
+//                    },
+//                    onClick = {
+//                        val intent = Intent(context, ExoPlayJet::class.java).apply {
+//                            putExtra("video_url", channel.channel_url ?: "")
+//                            putExtra("zone", "TV")
+//                            // Prepare channel list for ExoPlayJet
+//                            val allChannelsData = ArrayList((filteredChannels.value ?: emptyList()).mapNotNull { ch ->
+//                                if (ch == null) return@mapNotNull null
+//                                ChannelInfo(
+//                                    ch.channel_url ?: "",
+//                                    "http://localhost:$localPORT/jtvimage/${ch.logoUrl ?: ""}",
+//                                    ch.channel_name ?: ""
+//                                )
+//                            })
+//                            putParcelableArrayListExtra("channel_list_data", allChannelsData)
+//                            val currentChannelIndex = (filteredChannels.value ?: emptyList()).indexOf(channel)
+//                            putExtra("current_channel_index", currentChannelIndex)
+//                            putExtra("logo_url", "http://localhost:$localPORT/jtvimage/${channel.logoUrl ?: ""}")
+//                            putExtra("ch_name", channel.channel_name ?: "")
+//                        }
+//                        startActivity(context, intent, null)
+//
+//                        // Update recent channels
+//                        val recentChannelsJson = preferenceManager.myPrefs.recentChannels
+//                        val type = object : TypeToken<List<Channel>>() {}.type
+//                        val recentChannels: MutableList<Channel> =
+//                            if (!recentChannelsJson.isNullOrEmpty())
+//                                Gson().fromJson(recentChannelsJson, type) ?: mutableListOf()
+//                            else
+//                                mutableListOf()
+//
+//                        val existingIndex = recentChannels.indexOfFirst { it.channel_id == channel.channel_id }
+//                        if (existingIndex != -1) {
+//                            val existingChannel = recentChannels[existingIndex]
+//                            recentChannels.removeAt(existingIndex)
+//                            recentChannels.add(0, existingChannel)
+//                        } else {
+//                            recentChannels.add(0, channel)
+//                            if (recentChannels.size > 25) {
+//                                recentChannels.removeAt(recentChannels.size - 1)
+//                            }
+//                        }
+//
+//                        preferenceManager.myPrefs.recentChannels = Gson().toJson(recentChannels)
+//                        preferenceManager.savePreferences()
+//                    },
+//                    colors = androidx.tv.material3.CardDefaults.colors(
+//                        containerColor = MaterialTheme.colorScheme.primaryContainer,
+//                    )
+//                )
+//            }
+//        }
     }
 }
