+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/dev/Recent_Layout.kt
@@ -6,6 +6,7 @@ import android.util.Log
 import androidx.compose.animation.core.Spring
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.spring
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -16,6 +17,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.MaterialTheme
@@ -34,6 +36,10 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.draw.scale
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.graphics.Color
 import androidx.compose.material3.Text as CText
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
@@ -74,6 +80,7 @@ fun Recent_Layout(context: Context) {
                 modifier = Modifier.padding(top = 8.dp)
             )
         } else {
+            val focusRequester = remember { FocusRequester() }
             LazyVerticalGrid(
                 columns = GridCells.Adaptive(minSize = 100.dp),
                 contentPadding = PaddingValues(16.dp),
@@ -84,19 +91,15 @@ fun Recent_Layout(context: Context) {
                 items(recentChannels.value) { channel ->
 
                     var isFocused by remember { mutableStateOf(false) }
-                    val scale by animateFloatAsState(
-                        targetValue = if (isFocused) 1.1f else 1f,
-                        animationSpec = spring(
-                            dampingRatio = Spring.DampingRatioMediumBouncy,
-                            stiffness = Spring.StiffnessLow
-                        ), label = ""
-                    )
 
-                    ElevatedCard(
+                    Card(
                         elevation = CardDefaults.cardElevation(defaultElevation = 6.dp),
                         modifier = Modifier
                             .height(120.dp)
-                            .scale(scale)
+                            .focusRequester(focusRequester)
+                            .onFocusChanged { focusState ->
+                                isFocused = focusState.isFocused
+                            }
                             .clickable {
                                 Log.d("HT", channel.channel_name)
                                 val intent = Intent(context, ExoPlayJet::class.java).apply {
@@ -140,6 +143,7 @@ fun Recent_Layout(context: Context) {
                                 preferenceManager.myPrefs.recentChannels = recentChannelsJsonx
                                 preferenceManager.savePreferences()
                             },
+                        border = if (isFocused) BorderStroke(4.dp, Color(0xFFFFD700)) else null,
                         colors = CardDefaults.cardColors(
                             containerColor = MaterialTheme.colorScheme.outlineVariant,
                         )
