+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/dev/depreciated/Main_Layout_3rd_old.kt
@@ -1,18 +1,19 @@
-package com.skylake.skytv.jgorunner.ui.dev
+package com.skylake.skytv.jgorunner.ui.dev.depreciated
 
+import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
 import android.util.Log
 import androidx.compose.animation.core.Spring
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.spring
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Done
 import androidx.compose.material3.*
@@ -27,38 +28,26 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.core.content.ContextCompat
+import androidx.core.content.ContextCompat.startActivity
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import com.skylake.skytv.jgorunner.activities.ChannelInfo
 import com.skylake.skytv.jgorunner.data.SkySharedPref
-import androidx.tv.material3.CardDefaults as CardDefaultsTV
-import androidx.tv.material3.ClassicCard
-import androidx.compose.material3.Text
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontWeight
-import androidx.core.content.ContextCompat.startActivity
 import com.skylake.skytv.jgorunner.services.player.ExoPlayJet
+import com.skylake.skytv.jgorunner.ui.dev.M3UChannelExp
 import com.skylake.skytv.jgorunner.ui.screens.AppStartTracker
 
-@OptIn(ExperimentalGlideComposeApi::class)
+
+@SuppressLint("NewApi")
 @Composable
-fun Main_Layout_3rd(context: Context) {
+fun Main_Layout_3rd_old(context: Context) {
     val preferenceManager = remember { SkySharedPref.getInstance(context) }
     var allChannels by remember { mutableStateOf<List<M3UChannelExp>>(emptyList()) }
     var isLoading by remember { mutableStateOf(true) }
     var selectedCategory by remember { mutableStateOf<String?>(null) }
-
-    val localPORT by remember {
-        mutableIntStateOf(preferenceManager.myPrefs.jtvGoServerPort ?: 8080)
-    }
-    val basefinURL = "http://localhost:$localPORT"
-    var selectedChannel by remember { mutableStateOf<M3UChannelExp?>(null) }
-    var epgData by remember { mutableStateOf<EpgProgram?>(null) }
-    var epgDebugVar by remember { mutableStateOf(preferenceManager.myPrefs.epgDebug) }
 
     val categories = remember(allChannels) {
         listOf("All") + allChannels.mapNotNull { it.category }.distinct().sorted()
@@ -127,19 +116,6 @@ fun Main_Layout_3rd(context: Context) {
         }
     }
 
-    // Fetch EPG data for selected channel
-    LaunchedEffect(selectedChannel) {
-        selectedChannel?.let { channel ->
-            val chID = extractChannelIdFromPlayUrl(channel.url)
-            val epgURLc = "$basefinURL/epg/${chID ?: ""}/0"
-            Log.d("NANOdix3rd",epgURLc)
-            epgData = ChannelUtils.fetchEpg(epgURLc)
-        } ?: run {
-            epgData = null
-        }
-    }
-
-
 
     if (isLoading) {
         Box(contentAlignment = Alignment.Center, modifier = Modifier.fillMaxSize()) {
@@ -161,99 +137,10 @@ fun Main_Layout_3rd(context: Context) {
                         preferenceManager.savePreferences()
                     }
                 )
-                //////////
-
-                if (epgDebugVar) {
-                    if (epgData == null) {
-                        Box(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .height(120.dp)
-                                .padding(horizontal = 12.dp, vertical = 4.dp),
-                            contentAlignment = Alignment.Center
-                        ) {
-                            Text(
-                                text = "Select channel to show EPG",
-                                style = TextStyle(fontSize = 16.sp, fontWeight = FontWeight.Medium),
-                                color = Color.Gray
-                            )
-                        }
-                    } else {
-                        val epg = epgData
-                        Column(
-                            modifier = Modifier
-                                .padding(horizontal = 12.dp, vertical = 4.dp)
-                                .fillMaxWidth()
-                        ) {
-                            Card(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .height(120.dp),
-                                elevation = CardDefaults.cardElevation(defaultElevation = 6.dp),
-                                shape = RoundedCornerShape(16.dp)
-                            ) {
-                                Row(
-                                    modifier = Modifier
-                                        .fillMaxWidth()
-                                        .padding(8.dp),
-                                    verticalAlignment = Alignment.CenterVertically
-                                ) {
-                                    Column(
-                                        modifier = Modifier
-                                            .weight(1f)
-                                            .padding(start = 8.dp, end = 12.dp)
-                                    ) {
-                                        Text(
-                                            text = epg?.channel_name ?: "Unknown Channel",
-                                            style = TextStyle(fontSize = 14.sp)
-                                        )
-                                        Spacer(modifier = Modifier.height(6.dp))
-                                        Text(
-                                            text = epg?.showname ?: "No Program Info",
-                                            maxLines = 1,
-                                            style = TextStyle(fontSize = 22.sp, fontWeight = FontWeight.Bold)
-                                        )
-                                        Spacer(modifier = Modifier.height(8.dp))
-                                        Text(
-                                            text = epg?.description ?: "No description available",
-                                            style = TextStyle(fontSize = 13.sp),
-                                            maxLines = 3,
-                                            overflow = TextOverflow.Ellipsis
-                                        )
-                                    }
-                                    epg?.episodePoster?.let { poster ->
-                                        GlideImage(
-                                            model = "$basefinURL/jtvposter/$poster",
-                                            contentDescription = null,
-                                            modifier = Modifier
-                                                .height(90.dp)
-                                                .clip(RoundedCornerShape(12.dp)),
-                                            contentScale = ContentScale.Fit
-                                        )
-                                    }
-                                }
-                            }
-                        }
-                    }
-                } else {
-                    // Empty
-                }
-
-
-                ChannelGridTV(
+                ChannelGrid(
                     context = context,
-                    channels = filteredChannels,
-                    selectedChannel = selectedChannel,
-                    onSelectedChannelChanged = { channel -> selectedChannel = channel }
+                    channels = filteredChannels
                 )
-
-                ///////////
-//                ChannelGrid(
-//                    context = context,
-//                    channels = filteredChannels,
-//                    selectedChannel = selectedChannel,
-//                    onSelectedChannelChanged = { channel -> selectedChannel = channel }
-//                )
             }
         }
     }
@@ -290,12 +177,7 @@ private fun CategoryFilterRow(
 
 @OptIn(ExperimentalGlideComposeApi::class)
 @Composable
-private fun ChannelGrid(
-    context: Context,
-    channels: List<M3UChannelExp>,
-    selectedChannel: M3UChannelExp?,
-    onSelectedChannelChanged: (M3UChannelExp) -> Unit
-) {
+private fun ChannelGrid(context: Context, channels: List<M3UChannelExp>) {
     val focusRequester = remember { FocusRequester() }
     LazyVerticalGrid(
         columns = GridCells.Adaptive(minSize = 100.dp),
@@ -314,26 +196,35 @@ private fun ChannelGrid(
                 label = "scaleAnimation"
             )
 
-            ClassicCard(
+            ElevatedCard(
+                elevation = CardDefaults.cardElevation(defaultElevation = 6.dp),
                 modifier = Modifier
                     .height(120.dp)
                     .scale(scale)
                     .focusRequester(focusRequester)
-                    .onFocusChanged { focusState ->
-                        isFocused = focusState.isFocused
-                        if (focusState.isFocused) {
-                            onSelectedChannelChanged(channel)
+                    .onFocusChanged { focusState -> isFocused = focusState.isFocused }
+                    .clickable {
+                        Log.d("ChannelGrid", "Clicked on: ${channel.name} - ${channel.url}")
+                        val channelInfoList = ArrayList(channels.map {
+                            ChannelInfo(it.url, it.logo ?: "", it.name)
+                        })
+                        val currentIndex = channels.indexOf(channel)
+
+                        val intent = Intent(context, ExoPlayJet::class.java).apply {
+                            putParcelableArrayListExtra("channel_list_data", channelInfoList)
+                            putExtra("current_channel_index", currentIndex)
                         }
+                        ContextCompat.startActivity(context, intent, null)
                     },
-                image = {
+                colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.primaryContainer)
+            ) {
+                Column {
                     GlideImage(
                         model = channel.logo,
                         contentDescription = "${channel.name} logo",
                         modifier = Modifier.fillMaxWidth().height(80.dp),
                         contentScale = ContentScale.Fit
                     )
-                },
-                title = {
                     Text(
                         text = channel.name,
                         style = MaterialTheme.typography.bodyMedium,
@@ -342,24 +233,8 @@ private fun ChannelGrid(
                         fontSize = 12.sp,
                         modifier = Modifier.padding(8.dp)
                     )
-                },
-                onClick = {
-                    Log.d("ChannelGrid", "Clicked on: ${channel.name} - ${channel.url}")
-                    val channelInfoList = ArrayList(channels.map {
-                        ChannelInfo(it.url, it.logo ?: "", it.name)
-                    })
-                    val currentIndex = channels.indexOf(channel)
-
-                    val intent = Intent(context, ExoPlayJet::class.java).apply {
-                        putParcelableArrayListExtra("channel_list_data", channelInfoList)
-                        putExtra("current_channel_index", currentIndex)
-                    }
-                    startActivity(context, intent, null)
-                },
-                colors = CardDefaultsTV.colors(
-                    containerColor = MaterialTheme.colorScheme.primaryContainer,
-                )
-            )
+                }
+            }
         }
     }
 }
