+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/dev/depreciated/Main_Layout_old.kt
@@ -1,60 +1,60 @@
-package com.skylake.skytv.jgorunner.ui.dev
+package com.skylake.skytv.jgorunner.ui.dev.depreciated
 
+import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.util.Log
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.grid.GridCells
 import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Done
-import androidx.compose.material.icons.filled.Refresh
 import androidx.compose.material3.*
-import androidx.tv.material3.ClassicCard
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.core.content.ContextCompat.startActivity
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
-import com.skylake.skytv.jgorunner.activities.ChannelInfo
 import com.skylake.skytv.jgorunner.data.SkySharedPref
-import com.google.gson.Gson
-import com.google.gson.reflect.TypeToken
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.ElevatedCard
-import androidx.compose.material3.FilterChip
-import androidx.compose.material3.FilterChipDefaults
-import androidx.compose.material3.Icon
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.focus.onFocusEvent
 import androidx.compose.ui.layout.ContentScale
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Done
+import androidx.compose.material.icons.filled.Refresh
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
+import com.skylake.skytv.jgorunner.activities.ChannelInfo
 import com.skylake.skytv.jgorunner.services.player.ExoPlayJet
+import com.skylake.skytv.jgorunner.ui.dev.Channel
+import com.skylake.skytv.jgorunner.ui.dev.ChannelResponse
+import com.skylake.skytv.jgorunner.ui.dev.ChannelUtils
+import com.skylake.skytv.jgorunner.ui.dev.EpgProgram
 import com.skylake.skytv.jgorunner.ui.screens.AppStartTracker
 
+@SuppressLint("MutableCollectionMutableState")
 @OptIn(ExperimentalGlideComposeApi::class)
 @Composable
-fun Main_LayoutTV(context: Context?) {
+fun Main_Layout_old(context: Context?) {
     if (context == null) return
     val scope = rememberCoroutineScope()
     val channelsResponse = remember { mutableStateOf<ChannelResponse?>(null) }
     val filteredChannels = remember { mutableStateOf<List<Channel>>(emptyList()) }
     val preferenceManager = SkySharedPref.getInstance(context)
     val localPORT by remember {
-        mutableIntStateOf(preferenceManager.myPrefs.jtvGoServerPort ?: 8080)
+        mutableIntStateOf(preferenceManager.myPrefs.jtvGoServerPort)
     }
     val basefinURL = "http://localhost:$localPORT"
     var fetched by remember { mutableStateOf(false) }
@@ -63,6 +63,7 @@ fun Main_LayoutTV(context: Context?) {
     var epgData by remember { mutableStateOf<EpgProgram?>(null) }
 
     val focusRequester = remember { FocusRequester() }
+
     val categoryMap = mapOf(
         "Reset" to null,
         "Entertainment" to 5,
@@ -80,9 +81,11 @@ fun Main_LayoutTV(context: Context?) {
         "JioDarshan" to 19
     )
 
+    val categoryIdMap = categoryMap.filterValues { it != null }.map { it.value!! to it.key }.toMap()
     val savedCategoryIds = preferenceManager.myPrefs.filterCI
         ?.split(",")?.mapNotNull { it.toIntOrNull() }?.toMutableSet() ?: mutableSetOf()
     var selectedCategoryIds by remember { mutableStateOf(savedCategoryIds) }
+    val selectedCategories = categoryMap.filterValues { it in selectedCategoryIds }.keys
 
     val sortedCategories = remember(selectedCategoryIds) {
         val resetCategoryName = "Reset"
@@ -95,14 +98,13 @@ fun Main_LayoutTV(context: Context?) {
         listOf(resetCategoryName) + selectedOtherCategories + unselectedOtherCategories
     }
 
-    // Fetch and filter channels (cache/network)
     LaunchedEffect(Unit) {
         val sharedPref = context.getSharedPreferences("channel_cache", Context.MODE_PRIVATE)
         var useCache = true
         var cachedChannels: ChannelResponse? = null
 
         val cachedJson = sharedPref.getString("channels_json", null)
-        if (!cachedJson.isNullOrEmpty()) {
+        if (cachedJson != null) {
             try {
                 cachedChannels = Gson().fromJson(cachedJson, ChannelResponse::class.java)
                 channelsResponse.value = cachedChannels
@@ -122,22 +124,36 @@ fun Main_LayoutTV(context: Context?) {
                 ?.split(",")?.map { it.trim() }?.filter { it.isNotEmpty() }
             val languages = preferenceManager.myPrefs.filterLI
                 ?.split(",")?.map { it.trim() }?.filter { it.isNotEmpty() }
-
             val filtered = when {
-                categories.isNullOrEmpty() && languages.isNullOrEmpty() -> ChannelUtils.filterChannels(cachedChannels)
-                categories.isNullOrEmpty() -> ChannelUtils.filterChannels(
-                    cachedChannels,
-                    languageIds = languages?.mapNotNull { it.toIntOrNull() }?.takeIf { it.isNotEmpty() }
-                )
-                languages.isNullOrEmpty() -> ChannelUtils.filterChannels(
-                    cachedChannels,
-                    categoryIds = categories.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() }
-                )
-                else -> ChannelUtils.filterChannels(
-                    cachedChannels,
-                    categoryIds = categories.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() },
-                    languageIds = languages.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() }
-                )
+                categories.isNullOrEmpty() && languages.isNullOrEmpty() -> {
+                    ChannelUtils.filterChannels(cachedChannels)
+                }
+
+                categories.isNullOrEmpty() -> {
+                    ChannelUtils.filterChannels(
+                        cachedChannels,
+                        languageIds = languages?.mapNotNull { it.toIntOrNull() }
+                            ?.takeIf { it.isNotEmpty() }
+                    )
+                }
+
+                languages.isNullOrEmpty() -> {
+                    ChannelUtils.filterChannels(
+                        cachedChannels,
+                        categoryIds = categories.mapNotNull { it.toIntOrNull() }
+                            .takeIf { it.isNotEmpty() }
+                    )
+                }
+
+                else -> {
+                    ChannelUtils.filterChannels(
+                        cachedChannels,
+                        categoryIds = categories.mapNotNull { it.toIntOrNull() }
+                            .takeIf { it.isNotEmpty() },
+                        languageIds = languages.mapNotNull { it.toIntOrNull() }
+                            .takeIf { it.isNotEmpty() }
+                    )
+                }
             }
             filteredChannels.value = filtered ?: emptyList()
             fetched = true
@@ -149,6 +165,7 @@ fun Main_LayoutTV(context: Context?) {
                 try {
                     val response = ChannelUtils.fetchChannels("$basefinURL/channels")
                     channelsResponse.value = response
+
                     if (response != null) {
                         val responseJsonString = Gson().toJson(response)
                         with(sharedPref.edit()) {
@@ -159,28 +176,41 @@ fun Main_LayoutTV(context: Context?) {
                             ?.split(",")?.map { it.trim() }?.filter { it.isNotEmpty() }
                         val languages = preferenceManager.myPrefs.filterLI
                             ?.split(",")?.map { it.trim() }?.filter { it.isNotEmpty() }
-
                         val filtered = when {
-                            categories.isNullOrEmpty() && languages.isNullOrEmpty() -> ChannelUtils.filterChannels(response)
-                            categories.isNullOrEmpty() -> ChannelUtils.filterChannels(
-                                response,
-                                languageIds = languages?.mapNotNull { it.toIntOrNull() }?.takeIf { it.isNotEmpty() }
-                            )
-                            languages.isNullOrEmpty() -> ChannelUtils.filterChannels(
-                                response,
-                                categoryIds = categories.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() }
-                            )
-                            else -> ChannelUtils.filterChannels(
-                                response,
-                                categoryIds = categories.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() },
-                                languageIds = languages.mapNotNull { it.toIntOrNull() }.takeIf { it.isNotEmpty() }
-                            )
+                            categories.isNullOrEmpty() && languages.isNullOrEmpty() -> {
+                                ChannelUtils.filterChannels(response)
+                            }
+
+                            categories.isNullOrEmpty() -> {
+                                ChannelUtils.filterChannels(
+                                    response,
+                                    languageIds = languages?.mapNotNull { it.toIntOrNull() }
+                                        ?.takeIf { it.isNotEmpty() }
+                                )
+                            }
+
+                            languages.isNullOrEmpty() -> {
+                                ChannelUtils.filterChannels(
+                                    response,
+                                    categoryIds = categories.mapNotNull { it.toIntOrNull() }
+                                        ?.takeIf { it.isNotEmpty() }
+                                )
+                            }
+
+                            else -> {
+                                ChannelUtils.filterChannels(
+                                    response,
+                                    categoryIds = categories.mapNotNull { it.toIntOrNull() }
+                                        ?.takeIf { it.isNotEmpty() },
+                                    languageIds = languages.mapNotNull { it.toIntOrNull() }
+                                        ?.takeIf { it.isNotEmpty() }
+                                )
+                            }
                         }
                         filteredChannels.value = filtered ?: emptyList()
                         success = true
                     }
-                } catch (e: Exception) {
-                    // ignore, retry
+                } catch (_: Exception) {
                 }
                 if (!success) {
                     kotlinx.coroutines.delay(300)
@@ -188,6 +218,9 @@ fun Main_LayoutTV(context: Context?) {
             }
             fetched = true
         }
+
+
+
 
         if (preferenceManager.myPrefs.startTvAutomatically) {
             if (!AppStartTracker.shouldPlayChannel &&
@@ -214,6 +247,7 @@ fun Main_LayoutTV(context: Context?) {
                     putExtra("ch_name", firstChannel.channel_name ?: "")
                 }
                 kotlinx.coroutines.delay(1000)
+
                 startActivity(context, intent, null)
 
                 //Recent Channels
@@ -241,10 +275,12 @@ fun Main_LayoutTV(context: Context?) {
 
             AppStartTracker.shouldPlayChannel = true
         }
-
     }
 
-    // Re-filter channels when category changes
+
+
+
+
     LaunchedEffect(selectedCategoryIds) {
         channelsResponse.value?.let { response ->
             val languages = preferenceManager.myPrefs.filterLI
@@ -259,18 +295,17 @@ fun Main_LayoutTV(context: Context?) {
         }
     }
 
-    // Fetch EPG data for selected channel
     LaunchedEffect(selectedChannel) {
         selectedChannel?.let { channel ->
-            val epgURLc = "$basefinURL/epg/${channel.channel_id ?: ""}/0"
+            val epgURLc = "$basefinURL/epg/${channel.channel_id}/0"
             Log.d("NANOdix",epgURLc)
             epgData = ChannelUtils.fetchEpg(epgURLc)
+            Log.d("NANOdix", "Now playing: ${epgData?.showname}")
         } ?: run {
             epgData = null
         }
     }
 
-    // UI: Loading state
     if (!fetched) {
         Box(
             contentAlignment = Alignment.Center,
@@ -278,9 +313,7 @@ fun Main_LayoutTV(context: Context?) {
         ) {
             CircularProgressIndicator(modifier = Modifier.size(60.dp))
         }
-    }
-    // UI: Empty state
-    else if (filteredChannels.value.isNullOrEmpty()) {
+    } else if (filteredChannels.value.isNullOrEmpty()) {
         Box(
             contentAlignment = Alignment.Center,
             modifier = Modifier.fillMaxSize()
@@ -288,18 +321,18 @@ fun Main_LayoutTV(context: Context?) {
             Column(horizontalAlignment = Alignment.CenterHorizontally) {
                 Text(
                     text = "No channels found",
-                    style = TextStyle(fontSize = 20.sp, fontWeight = FontWeight.Bold),
+                    style = MaterialTheme.typography.titleLarge,
                     color = Color.Red
                 )
                 Spacer(modifier = Modifier.height(16.dp))
                 Text(
                     text = "Try the following steps:",
-                    style = TextStyle(fontSize = 16.sp)
+                    style = MaterialTheme.typography.bodyLarge
                 )
                 Spacer(modifier = Modifier.height(8.dp))
                 Text(
                     text = "\n• Check your internet connection\n• Reload the app\n• Go to Main Screen for detailed information",
-                    style = TextStyle(fontSize = 15.sp, color = Color.Gray)
+                    style = MaterialTheme.typography.bodyMedium.copy(color = Color.Gray)
                 )
                 Spacer(modifier = Modifier.height(24.dp))
                 ElevatedCard(
@@ -316,19 +349,14 @@ fun Main_LayoutTV(context: Context?) {
                         verticalAlignment = Alignment.CenterVertically,
                         modifier = Modifier.padding(horizontal = 24.dp, vertical = 12.dp)
                     ) {
-                        Icon(
-                            imageVector = Icons.Filled.Refresh,
-                            contentDescription = "Reload"
-                        )
+                        Icon(imageVector = Icons.Filled.Refresh, contentDescription = "Reload")
                         Spacer(modifier = Modifier.width(8.dp))
                         Text("Reload App")
                     }
                 }
             }
         }
-    }
-    // UI: Main content
-    else {
+    } else {
         // CATEGORY CHIPS
         LazyRow(
             modifier = Modifier
@@ -339,7 +367,6 @@ fun Main_LayoutTV(context: Context?) {
             items(sortedCategories) { categoryName ->
                 val categoryId = categoryMap[categoryName]
                 val isSelected = categoryId != null && selectedCategoryIds.contains(categoryId)
-
                 FilterChip(
                     onClick = {
                         if (categoryName == "Reset") {
@@ -356,7 +383,7 @@ fun Main_LayoutTV(context: Context?) {
                         preferenceManager.myPrefs.filterCI = updatedCI
                         preferenceManager.savePreferences()
                     },
-                    label = { Text(categoryName ?: "Unknown") },
+                    label = { Text(categoryName) },
                     selected = isSelected,
                     leadingIcon = when {
                         categoryName == "Reset" -> {
@@ -383,7 +410,7 @@ fun Main_LayoutTV(context: Context?) {
             }
         }
 
-        // EPG CARD (null-safe)
+        // EPG CARD
         epgData?.let { epg ->
             Column(
                 modifier = Modifier
@@ -408,22 +435,23 @@ fun Main_LayoutTV(context: Context?) {
                         ) {
                             Text(
                                 text = epg.channel_name ?: "Unknown Channel",
-                                style = TextStyle(fontSize = 14.sp)
+                                style = MaterialTheme.typography.bodyMedium
                             )
                             Spacer(modifier = Modifier.height(6.dp))
                             Text(
                                 text = epg.showname ?: "No Program Info",
                                 maxLines = 1,
-                                style = TextStyle(fontSize = 22.sp, fontWeight = FontWeight.Bold)
+                                style = MaterialTheme.typography.titleLarge.copy(fontWeight = FontWeight.Bold)
                             )
                             Spacer(modifier = Modifier.height(8.dp))
                             Text(
                                 text = epg.description ?: "No description available",
-                                style = TextStyle(fontSize = 13.sp),
-                                maxLines = 3,
+                                style = MaterialTheme.typography.bodySmall,
+                                maxLines = 2,
                                 overflow = TextOverflow.Ellipsis
                             )
                         }
+
                         epg.episodePoster?.let { poster ->
                             GlideImage(
                                 model = "$basefinURL/jtvposter/$poster",
@@ -447,83 +475,86 @@ fun Main_LayoutTV(context: Context?) {
             horizontalArrangement = Arrangement.spacedBy(16.dp)
         ) {
             items(filteredChannels.value ?: emptyList()) { channel ->
-                if (channel == null) return@items // Defensive: skip nulls
-                ClassicCard(
+                var isFocused by remember { mutableStateOf(false) }
+                ElevatedCard(
+                    elevation = CardDefaults.cardElevation(defaultElevation = 6.dp),
                     modifier = Modifier
                         .height(120.dp)
                         .focusRequester(focusRequester)
                         .onFocusEvent { focusState ->
+                            isFocused = focusState.isFocused
                             if (focusState.isFocused) {
                                 selectedChannel = channel
                             }
+                        }
+                        .clickable {
+                            val intent = Intent(context, ExoPlayJet::class.java).apply {
+                                putExtra("zone", "TV")
+                                putParcelableArrayListExtra("channel_list_data", ArrayList(
+                                    (filteredChannels.value ?: emptyList()).map { ch ->
+                                        ChannelInfo(
+                                            ch.channel_url ?: "",
+                                            "http://localhost:$localPORT/jtvimage/${ch.logoUrl ?: ""}",
+                                            ch.channel_name ?: ""
+                                        )
+                                    }
+                                ))
+                                putExtra("current_channel_index", (filteredChannels.value ?: emptyList()).indexOf(channel))
+                                putExtra("video_url", channel.channel_url ?: "")
+                                putExtra("logo_url", "http://localhost:$localPORT/jtvimage/${channel.logoUrl ?: ""}")
+                                putExtra("ch_name", channel.channel_name ?: "")
+                            }
+                            startActivity(context, intent, null)
+
+                            // Update recent channels
+                            val recentChannelsJson = preferenceManager.myPrefs.recentChannels
+                            val type = object : TypeToken<List<Channel>>() {}.type
+                            val recentChannels: MutableList<Channel> =
+                                Gson().fromJson(recentChannelsJson, type) ?: mutableListOf()
+
+                            val existingIndex = recentChannels.indexOfFirst {
+                                it.channel_id == channel.channel_id
+                            }
+
+                            if (existingIndex != -1) {
+                                val existingChannel = recentChannels[existingIndex]
+                                recentChannels.removeAt(existingIndex)
+                                recentChannels.add(0, existingChannel)
+                            } else {
+                                recentChannels.add(0, channel)
+                                if (recentChannels.size > 25) {
+                                    recentChannels.removeAt(recentChannels.size - 1)
+                                }
+                            }
+
+                            preferenceManager.myPrefs.recentChannels = Gson().toJson(recentChannels)
+                            preferenceManager.savePreferences()
                         },
-                    image = {
-                        val logoUrl = channel.logoUrl ?: ""
-                        val imageUrl = "$basefinURL/jtvimage/$logoUrl"
+                    colors = CardDefaults.cardColors(
+                        containerColor = MaterialTheme.colorScheme.primaryContainer
+                    )
+                ) {
+                    Column {
+                        // Channel logo
+                        val imageUrl = "http://localhost:$localPORT/jtvimage/${channel.logoUrl ?: ""}"
+
                         GlideImage(
                             model = imageUrl,
-                            contentDescription = channel.channel_name ?: "Channel",
+                            contentDescription = channel.channel_name ?: "",
                             modifier = Modifier
                                 .fillMaxWidth()
-                                .height(80.dp)
+                                .height(80.dp),
+                            contentScale = ContentScale.Fit
                         )
-                    },
-                    title = {
+
+                        // Channel name
                         Text(
-                            text = channel.channel_name ?: "Channel",
+                            text = channel.channel_name ?: "Unknown",
                             fontSize = 12.sp,
                             modifier = Modifier.padding(8.dp)
                         )
-                    },
-                    onClick = {
-                        val intent = Intent(context, ExoPlayJet::class.java).apply {
-                            putExtra("video_url", channel.channel_url ?: "")
-                            putExtra("zone", "TV")
-                            // Prepare channel list for ExoPlayJet
-                            val allChannelsData = ArrayList((filteredChannels.value ?: emptyList()).mapNotNull { ch ->
-                                if (ch == null) return@mapNotNull null
-                                ChannelInfo(
-                                    ch.channel_url ?: "",
-                                    "http://localhost:$localPORT/jtvimage/${ch.logoUrl ?: ""}",
-                                    ch.channel_name ?: ""
-                                )
-                            })
-                            putParcelableArrayListExtra("channel_list_data", allChannelsData)
-                            val currentChannelIndex = (filteredChannels.value ?: emptyList()).indexOf(channel)
-                            putExtra("current_channel_index", currentChannelIndex)
-                            putExtra("logo_url", "http://localhost:$localPORT/jtvimage/${channel.logoUrl ?: ""}")
-                            putExtra("ch_name", channel.channel_name ?: "")
-                        }
-                        startActivity(context, intent, null)
-
-                        // Update recent channels
-                        val recentChannelsJson = preferenceManager.myPrefs.recentChannels
-                        val type = object : TypeToken<List<Channel>>() {}.type
-                        val recentChannels: MutableList<Channel> =
-                            if (!recentChannelsJson.isNullOrEmpty())
-                                Gson().fromJson(recentChannelsJson, type) ?: mutableListOf()
-                            else
-                                mutableListOf()
-
-                        val existingIndex = recentChannels.indexOfFirst { it.channel_id == channel.channel_id }
-                        if (existingIndex != -1) {
-                            val existingChannel = recentChannels[existingIndex]
-                            recentChannels.removeAt(existingIndex)
-                            recentChannels.add(0, existingChannel)
-                        } else {
-                            recentChannels.add(0, channel)
-                            if (recentChannels.size > 25) {
-                                recentChannels.removeAt(recentChannels.size - 1)
-                            }
-                        }
-
-                        preferenceManager.myPrefs.recentChannels = Gson().toJson(recentChannels)
-                        preferenceManager.savePreferences()
-                    },
-                    colors = androidx.tv.material3.CardDefaults.colors(
-                        containerColor = MaterialTheme.colorScheme.primaryContainer,
-                    )
-                )
+                    }
+                }
             }
         }
     }
