+++ app/src/main/java/com/skylake/skytv/jgorunner/ui/dev/depreciated/SearchTabLayout_TV.kt
@@ -1,4 +1,4 @@
-package com.skylake.skytv.jgorunner.ui.dev
+package com.skylake.skytv.jgorunner.ui.dev.depreciated
 
 import android.content.Context
 import android.content.Intent
@@ -23,9 +23,11 @@ import androidx.tv.material3.ClassicCard
 import com.bumptech.glide.integration.compose.GlideImage
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
-import com.skylake.skytv.jgorunner.activities.ExoplayerActivity
 import com.skylake.skytv.jgorunner.data.SkySharedPref
 import com.skylake.skytv.jgorunner.services.player.ExoPlayJet
+import com.skylake.skytv.jgorunner.ui.dev.Channel
+import com.skylake.skytv.jgorunner.ui.dev.ChannelResponse
+import com.skylake.skytv.jgorunner.ui.dev.ChannelUtils
 import kotlinx.coroutines.delay
 
 @OptIn(ExperimentalGlideComposeApi::class)
